refactor: remove express concepts from ui utils
